requirements.txt
annotated-types==0.7.0
anyio==4.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
exceptiongroup==1.2.2
fastapi==0.115.2
h11==0.14.0
idna==3.10
numpy==1.24.4
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
requests==2.32.3
sniffio==1.3.1
starlette==0.40.0
tinycss2==1.3.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
webencodings==0.5.1
